Split autoRefreshServer into its own file
